Allow big AS numbers in BGP peering sessions
Because:
- RFC 6793 extends the BGP protocol to allow for 4-byte (unsigned) AS
  numbers, as opposed to the old 2-byte numbers.
- The BGP peering session table in PostgreSQL uses INT data types to
  store the AS numbers, but these are signed, and will therefore only
  allow a maximum AS of 2,147,483,647, not 4,294,967,295 as the new
  standard specifies.

Fixes #2216

+++ python/nav/models/manage.py
@@ -2166,8 +2166,8 @@ class GatewayPeerSession(models.Model):
     protocol = models.IntegerField(choices=PROTOCOL_CHOICES)
     peer = models.GenericIPAddressField()
     state = VarcharField()
-    local_as = models.IntegerField(null=True)
-    remote_as = models.IntegerField(null=True)
+    local_as = models.BigIntegerField(null=True)
+    remote_as = models.BigIntegerField(null=True)
     adminstatus = VarcharField()
 
     class Meta(object):
